requirements.txt
strawberry-graphql
fastapi